refactor: timestamp based snapshots (#14128)
Fixes #13877.

Updates the address snapshot library to be based on time only, and rely
on the caller to pass meaningful time in. The staking lib can then feed
along the timestamp for which the current epoch is stable `timeOfEpoch -
1` and we can neatly use the same storage later for different rollups.

Alters quite a lot of the tests to test the timings.