feat: hide commitments and evaluations in (most of) the wires in translator (#14797)
Yay, translator is a bit more ZK!

Add random elements in Translator wire polynomials that are not
corresponding to the op queue and logic to transfer random values from
the interleaved polynomials to the first four ordered range constraints.
Disabling of random rows in translator for sumcheck is achieved via
relations, using a variety of `lagrange_*` polynomials so some relations
have been modified accordingly.