feat: slash proposers of invalid state root (#14791)
Turns the `ValidatorClient` into a `Watcher` that emits `WANT_TO_SLASH`
events whenever re-execution fails due to invalid state roots.

Within the `ValidatorClient`, maintain a `Set` of proposers who have
proposed invalid blocks such that the slasher client may come along
afterward and ask if the proposer should be slashed.

Add the `ValidatorClient` we create in `createAndSync` to the watchers
that the `SlasherClient` responds to and consults.

Note: this means that the `ValidatorClient` will be started/stopped
twice, but the operations are idempotent.

Fix #14409