fix: `epochs_monitor_block_building.test.ts` flake (#15497)
Facundo reported [this flake](http://ci.aztec-labs.com/da94e250fc8867e3)
and I am fixing it in this PR.

The cause of the issue was that in the test we had `minTxsPerBlock` set
to 0 and we were deploying 1 account and that led to an edge case. In
the setup we try to controllably progress past genesis by either
deploying an account in block 1 or waiting for an empty block to be
mined. Now the edge case was that we were trying to deploy the account
in block 1 but since initially the `minTxsPerBlock` was set to 0 the
chain has already progressed past genesis before the block with account
contract deployment was mined.

The fix is to set the modified `minTxsPerBlock` before the sequencer is
event started. That way we don't get the unexpectedly built empty block.