test: add public context txe + state var tests (#15706)
This does a couple things, all closely related:
- it enables the `PublicImmutable` tests, which were included in #6985
but never run, with some rewriting using the new features
 - it adds `PublicMutable` tests
 - it adds `env.public_context` tests
- it makes `public_context` smarter by being able to combine the
committed and transient state for nullifiers in a pub context, allowing
us to read transient nullifiers. This is avm behavior that we need to
replicate in txe, but I don't think there'll be too much more of it.