feat!: improve txe deploy API (#16011)
Built on top of
https://github.com/AztecProtocol/aztec-packages/pull/16007 as I'm
reusing the counters from there.

We can now have multiple deployments of the same contract class since
we'll assign them different secrets (we also could've specified a salt,
but this also works _and_ it gives us unique secrets, so it seems more
convenient). I also merged `deploy` and `deploy_self` by using `path` a
bit more to parse the string - we'll need to make this more robust in
the future to avoid having TXE crash when it doesn't find the file and
instead return a reasonable error, but it works for now. I also changed
the return value to be the address, since we never care about the full
instance (and we can retrieve it if needed).

---------

Co-authored-by: benesjan <janbenes1234@gmail.com>