chore(docs): more cleanup (#16442)
# Changes

- Removes an unneeded path in `tutorials/codealong`, fixes all those
links
- Removes an entire tutorial "Node.js app that interacts with contracts"
which was exactly the same as "Getting started with Aztec.js". The only
difference was that the former copy-pasted the Token contract, whereas
the latter imported it from `@aztec/noir-contracts.js`. Insane.

As a replacement:
- Adds a tiny tutorial aptly named `Deploy a Token Contract` which does
the same thing. No #include_code code snippets, this may backfire but it
would be easy enough to make the tutorial testable (like we do for
Barretenberg docs).
- Spins off a how-to guide on how to compile and deploy a random
contract

I think we can benefit from not leaning too heavily on code snippets.
Both `aztec.nr` and `aztec.js` are much stabler now and actually can't
change much as we rush towards mainnet. Since our #1 prio is fixing
docs, this will make us move faster.