feat: Slash lists, store expiration, veto checks (#16667)
Adds some missing features to slashing:

- Adds support for the slashValidatorNever and slashValidatorAlways
config settings (now EthAddress lists) so that validators in those lists
are never/always slashed. All validators in the local keystore are
automatically added to the "never" list, unless a `slashSelfAllowed` is
set.

- Checks if a slash payload is vetoed before trying to execute it, to
avoid running an unnecessary simulation. This is handled on the slasher
client directly.

- Adds expiration for offenses and payloads to avoid cluttering the
local stores. Removes unused `slashPayloadTtl setting`.

- Avoids posting committee addresses during executeRound for committees
that are not slashed, in order to save calldata gas.

- Moves slasher factory methods around to avoid circular dependency.

Previously posted as #16643 but was accidentally merged against the
wrong branch