feat(avm): public data squashing sorting (#17208)
This PR introduces a method which is called at the end of simulation in
order to generate `ff_gt` events which are required by the sub-trace
`public_data_squash` to sort the entries by `leaf_slot` (finite field
elements).

The required `ff_gt` events are generated based on the "public data
writes" elements part of the public inputs. The leaf_slots of these
"public data writes" are sorted in increasing order and an `ff_gt` event
contains two consecutive leaf_slots. It works because there is a
one-to-one correspondence between each different 'leaf_slot' in the
`public_data_squash` sub-trace and the "public data writes".

- public_inputs reference is passed to `AvmSimulationHelper` constructor
to initialize the required members (`public_data_writes_length` and
`public_data_writes`).
- New routine (`generate_ff_gt_events_for_squashing`) to emit `ff_gt`
events is defined in class `PublicDataTreeCheck`
- `AvmSimulationHelper` calls this routine after the main "tx simulation
call" ( `tx_execution.simulate(hints.tx)`)