refactor: remove typed oracle and txe typed oracle (#17279)
This PR removes `TypedOracle` and `TXETypedOracle`. In their place, it
creates 5 interfaces for the different kinds of oracles:
 - utility
 - private
 - avm
 - txe
- misc (a new kind, used for stuff that's available both in utility and
also top level tests, such as random())

These interfaces are then implemented by the different classes that
handle execution. The interfaces include runtime markers so that we can
figure out whether a generic handler implements a specific interface or
not, and throw during foreign call translation if the current handler
doesn't support it (e.g. when calling a private oracle in the top level
context in a test, or when calling a private oracle while simulating a
utility function). We used to be able to always call these functions and
then have their implementation error out via the default `TypedOracle`
implementation, but this was a bit messy as there was no actual way to
check for full interface implementation.

In doing this I found some minor issues with the oracles, which already
make this be valuable, e.g. there were some abandoned oracles that were
just never invoked but still declared in the TXETypedOracle object. I'll
do some oracle cleanup later on, but not now in order to try to make all
the changes happen in one go and reduce noise there.

The **real** reason why I made this change is because now `txe_oracle`
can be replaced by `private_execution_oracle` and
`utility_execution_oracle`, since they now implement the same interface.
This will mean that Noir tests will be using the real PXE implementation
and not a copy of it. I'll do this in a separate PR because there'll
likely be things that need to be adjusted there.

The diff is a bit messy due to interface changes combined with files
being renamed. I did not actually change any of the oracle methods.