chore: Deflake gossip network test (#17314)
As part of #17273 I had added a cleanup to the gossip network test to
delete data dirs for the prover. However, the `stop` method on the
prover failed to await for all operations, so when the test finished
successfully, it would still try to use the db (in particular, it seems
to be for the proving broker database `getEpochDatabase`) and abort with
a core dump.

This reverts the folder cleanup.