feat(avm)!: Calldata Hashing (#17224)
## Constraining Calldata Hashing

For a detailed overview + diagrams look at the design document! Here is
a brief-ish overview. Following a similar path to bytecode hashing
(#16756) , we use the `poseidon2` trace to prove correctness of calldata
hash and the calldata size.

When we initialise a `context` with a new public call request, we emit a
`CalldataEvent`:

```C++
struct CalldataEvent {
    uint32_t context_id;
    uint32_t calldata_size;
    std::vector<FF> calldata;
    // FF output_hash; // NOTE - removed as it was unused, should we emit this as well as a hint?
};
```

This data makes its way to `calldata.pil` for retrieval (from which
`data_copy.pil` extracts the fields) and `calldata_hashing.pil` for
hashing and size constraining. Each field in `calldata` is looked up and
matched to a field in `poseidon2`. At the final (`latch`) row, we
constrain that all calldata fields have been included (no more, no
less), the `output_hash` matches the `output` of the posiedon hash, and
the given `calldata_size` is correct.

This PR adds a new lookup from `tx.pil` to `calldata_hashing.pil`
linking the hash and size from the corresponding public call request.
Now, the existing lookup to the **hash** in the public inputs is fully
constrained to be correct.

It also adds `calldata_size` as a new column (and a new property in the
`EnqueuedCallEvent`) and appends it to the existing lookup to
`execution.pil` at dispatch, constraining that
`execution.parent_calldata_size` (taken from `context.pil`) matches.
This value is also looked up by `data_copy`, and so is now fully
constrained.

### Main TODOs/Open Qs:

- [ ] ~Check e2e tests~ + create test with explicit multiple calldata
reads
- [ ] Finalise approach of `read_addr_plus_one`
- [x] Finalise removal of hash from calldata event
- [ ] How best to combine some relations in `calldata_hashing` - i.e.
`#[START_AFTER_LATCH]` and ` #[START_INDEX_IS_ZERO]`?
`#[INDEX_INCREMENTS_i]`?