fix: deriving serde on more public structs (#17380)
In a PR down the stack I needed to slap `#[derive(Deserialize)]` on
`RetrievedNote` because I returned it from a function. Nico proposed
that we derive serde on more public structs because external won't be
able to do so.

In this PR I went through the public structs in Aztec.nr and applied
serde where I thought it could make sense to return from a contract
function - in most of the cases it didn't make sense.