chore: dependabot deps (#17418)
This PR updates most of the `yarn-project` dependabot alerted deps.

### On resolutions
I have tried to update all deps "properly", using `resolutions` as a
fallback. This is when we depend on D, which has outdated D' for which
Dependabot complains, and D doesn't have a newer version.

An example of this is
[docosaurus-plugin-ideal-image](https://github.com/facebook/docusaurus/blob/main/packages/docusaurus-plugin-ideal-image/package.json)
> sharp (outdated) > tar-fs (outdated) - meaning that
`docusaurus-plugin-ideal-image` imports an outdated version of `sharp`,
which in turn imports an outdated version of `tar-fs`, so we have to
force this fix via `resolutions`

### Not updated deps

I could not update two dependencies because the author has not fixed the
bug and published an update:
1. [bigint-buffer](https://github.com/no2chem/bigint-buffer/releases)
2. [private-ip](https://github.com/frenchbread/private-ip/releases)

Both of these are transitive deps, further complicating the update.