feat!: Databus consistency checks in CIVC verification (#17559)
We perform the databus consistency check in the ClientIVC recursive
verifier. The Hiding Kernel takes the return data of the tail kernel and
propagates it via its public inputs. Then, the ClientIVC recursive
verifier checks the calldata of the Hiding Kernel against the data in
the public inputs, thus enforcing that the Hiding Kernel used the data
from the Tail Kernel .

Closes https://github.com/AztecProtocol/barretenberg/issues/1510

---------

Co-authored-by: Khashayar Barooti <khashayar@aztecprotocol.com>