refactor!: dropping pub context mutable refs (#17997)
Closes
https://linear.app/aztec-labs/issue/F-110/remote-mut-refs-to-public-context

`PrivateContext` generally needs to be passed as a mutable reference to
functions because it does actually hold state
we're mutating. This is not the case for `PublicContext`, or
`UtilityContext` - these are just marker objects that
indicate the current execution mode and make available the correct
subset of the API. For this reason we have dropped
the mutable reference from the API.

Marked as breaking change because a dev needs to manually update the
type in case he passed the context around to custom functions.