refactor: dropping recipient from `NoteDao` and `UniqueNote` (#18586)
Fixes F-92

The issue description describes the motivation for this well:

> NoteDao contains a recipient value. This is no longer used as there is
no filtering by recipient. It is also incorrect as notes don't
technically have a recipient, they're just data. It is the note message
that has a recipient.
>
> As recipient gets removed from the object, we'll no longer be able to
use it to determine the note's scope. We should therefore ask the caller
to specify the scope, which they can just do with the recipient field
they used when creating the NoteDao (the message recipient).