chore: ACIR audit - part 7 (#18653)
### 🧾 Audit Context

This is the seventh PR in the thread of auditing ACIR. It is concerned
with refactoring with the goals of:
 1. removing reliance on `varnum` for write_vk scenarios
2. ensure the maximum witness index produced by ACIR isn't larger than
the length of the witness vector supplied
 3. improve code readability

### 🛠️ Changes Made

This PR:
- Removes `varnum` (used for ACIR generation) so that VKs can be
generated without the prior knowledge of how many witnesses a circuit
has.
- Introduces checks to ensure that the max witness index passed by ACIR
is not above length of the witness vector.
- Renames `has_dummy_witnesses` to `is_write_vk_mode` (which is more
descriptive)
- It refactors some test to make them more readable.
- It removes some tests which have been superseded.
- Restructures the code in `acir_to_constraint_buf` so that it is easier
to read.

**Note:** The removal of `varnum` can be discusssed. Instead of removing
it we could check its validity against the maximum witness index
encountered. I believe the calculation of the max witness index is
important as:
- it removes from barretenberg the requirement to know in advance how
many witnesses a builder will have
- it allows us to check that barretenberg was provided as many witnesses
as the ACIR constraints want to use

#### Background on varnum

When we build a circuit from ACIR we get the opcodes and a variable
called `varnum` (extracted from the circuit serialisation, where it is
called `current_witness_index`). The meaning of `varnum` is _the number
of witnesses known at the time of acir generation_. This number will in
general be smaller than the final number of witnesses in a circuit as
the latter depends on how certain gates are implemented.

When a circuit is built from ACIR together with a satisfying witness, we
always have `witness.size() = varnum = current_witness_index + 1`. This
is true because even if Noir optimises away some witnesses, barretenberg
fills the holes in the witness vector with zeros, so that the final
sizes match.

When we are in a write vk scenario, the witness vector is empty, so the
equality is not true anymore. This would not be a problem if it weren't
for the fact that barretenberg requires a builder to have the correct
number of variables even when we are writing a vk.

#### Removing varnum

To remove `varnum` we need to extract from the constraints the
`current_witness_index`, which is just the maximum witness index. To do
this, we keep a running maximum witness index and we update it every
time we encounter a new Acir::Opcode. We need to take into account even
Brillig calls to update the running maximum because even though Brillig
calls don't add constraints, they add witness values.

### ✅ Checklist

- [ ] Audited all methods of the relevant module/class
- [ ] Audited the interface of the module/class with other (relevant)
components
- [ ] Documented existing functionality and any changes made (as per
Doxygen requirements)
- [ ] Resolved and/or closed all issues/TODOs pertaining to the audited
files
- [ ] Confirmed and documented any security or other issues found (if
applicable)
- [ ] Verified that tests cover all critical paths (and added tests if
necessary)
- [ ] Updated audit tracking for the files audited (check the start of
each file you audited)

### 📌 Notes for Reviewers

(Optional) Call out anything that reviewers should pay close attention
to — like logic changes, performance implications, or potential
regressions.