chore: ACIR audit - part 8 (#18692)
### 🧾 Audit Context

This is the eight PR in the thread of auditing ACIR. This PR slightly
modifies the testing framework for DSL opcodes so that its flow is
closer that a real one.

### 🛠️ Changes Made

The testing classes `TestClass` and `TestClassPredicate` take a
constraint (one of the various constraints in stored in `AcirFormat`),
put it into an `AcirProgram` and pass this program to `create_circuit`.
This is close to what would happen if we were given the byte
serialisation of a Noir program, but it starts one level lower than it
should. The real flow for a Noir program is: Noir --> ACIR --> Bytes -->
`circuit_buf_to_acir_format` --> `create_circuit`. Our testing flow is
currently not testing `circuit_buf_to_acir_format`.

This PR modifies the testing flow as follows: given an ACIR constraint
in barretenberg, we reconstruct the `Acir::Opcode` from which such
constraint should have arisen. Then, this `Acir::Opcode` is put into an
`Acir::Circuit`, which is passed to `circuit_buf_to_acir_format`. From
here on, the testing flow stays the same.

The conversion from an ACIR constraint in barretenberg to an
`Acir::Opcode` is just a series of if/else to convert a list of
`uint32_t`/`WitnessOrConstant` into
`Acir::Witness`/`Acir::FunctionInput`.

**Note:** The new flow highlighted a deserialisation problem in
`circuit_buf_to_acir_format` that the old testing flow didn't catch.
This was not caught from the acir tests because Noir currently passes
witnesses in the place where the error was, but in barretenberg we are
meant to test both for witnesses and constants. Barretenberg was failing
when encountering constants.

We also add docs for `acir_format`.

### ✅ Checklist

- [ ] Audited all methods of the relevant module/class
- [ ] Audited the interface of the module/class with other (relevant)
components
- [ ] Documented existing functionality and any changes made (as per
Doxygen requirements)
- [ ] Resolved and/or closed all issues/TODOs pertaining to the audited
files
- [ ] Confirmed and documented any security or other issues found (if
applicable)
- [ ] Verified that tests cover all critical paths (and added tests if
necessary)
- [ ] Updated audit tracking for the files audited (check the start of
each file you audited)

### 📌 Notes for Reviewers

(Optional) Call out anything that reviewers should pay close attention
to — like logic changes, performance implications, or potential
regressions.