fix(avm): the "deepest" error for C++ simulations is the LAST one at the deepest level (#18959)
Prior to this, if there are multiple errors at the same level,
`findDeepestRevert` would return the FIRST (earliest) one. This was
different from TS, and intuitively I think it should be the LAST (most
recent).

This was encountered when running opcode spam tests. Example:
1. Make nested call
2. run EMITNOTEHASH up until limit (but not past limit)
3. explicit REVERT opcode
4. Make nested call again

This REVERT happens like 1000 times until eventually the nested call
actually halts due to "out of gas". That is the last error at that
deepest level.

- TS reported "out of gas"
- C++ reported "Assertion failed" (aka REVERT opcode reached)